Widen ABI fixed-array length to u64 and raise canonical type-layout bounds
- Change TypeRef::Array length from u32 to u64 in psy_data ABI and
  psy_compiler ABI, with u64::MAX fallback on conversion
- Raise canonical layout limits: MAX_ARRAY_LENGTH = 2^32,
  MAX_SLOT_COUNT = 2^34, and widen circuit range checks accordingly
  (31→33 bits for args, 32→35 bits for totals)
- Regenerate cached circuit library
- Point client_prover Makefile deploy/update targets at
  psy-genesis token artifacts via $(CURDIR)/../psy-genesis

+++ client_prover/psy_compiler/src/abi/mod.rs
@@ -200,7 +200,7 @@ pub enum TypeRef {
     },
     Array {
         item: Box<TypeRef>,
-        length: u32,
+        length: u64,
         item_felt_size: usize,
     },
     Map {
@@ -556,7 +556,7 @@ impl TypeRef {
                 let item_felt_size = element.felt_size(struct_layouts).unwrap_or(0);
                 TypeRef::Array {
                     item,
-                    length: (*count).try_into().unwrap_or(u32::MAX),
+                    length: (*count).try_into().unwrap_or(u64::MAX),
                     item_felt_size,
                 }
             }
@@ -565,7 +565,7 @@ impl TypeRef {
                 let item_felt_size = element.felt_size(struct_layouts).unwrap_or(0);
                 TypeRef::Array {
                     item,
-                    length: (*count).try_into().unwrap_or(u32::MAX),
+                    length: (*count).try_into().unwrap_or(u64::MAX),
                     item_felt_size,
                 }
             }
